Remove unused `Haddock.Utils` functions
  * removed functions in `Haddock.Utils` that were not used anywhere
    (or exported from the `haddock-api` package)

  * moved GHC-specific utils from `Haddock.Utils` to `Haddock.GhcUtils`
